requirements.txt
fastapi
assemblyai
python-decouple
requests
moviepy
uvicorn
